+++ app/helpers/legal_panel_for_government/journey_helper.rb
@@ -7,15 +7,6 @@ module LegalPanelForGovernment::JourneyHelper
     t('legal_panel_for_government.journey.lot_legal_specialisms', lot_number: lot.number)
   end
 
-  def download_legal_panel_for_government_suppliers_path
-    download_legal_panel_for_government_rm6360_suppliers_path(
-      lot_id: params[:lot_id],
-      service_ids: params[:service_ids],
-      not_core_jurisdiction: params[:not_core_jurisdiction],
-      jurisdiction_ids: params[:jurisdiction_ids]
-    )
-  end
-
   def prospectus_link_present?(supplier_framework, lot)
     prospectus_link(supplier_framework, lot).present? && prospectus_link(supplier_framework, lot).downcase != 'n/a'
   end