+++ app/services/legal_panel_for_government/rm6360/supplier_spreadsheet_creator.rb
@@ -1,14 +1,45 @@
 class LegalPanelForGovernment::RM6360::SupplierSpreadsheetCreator < SupplierSpreadsheetCreator
-  def build
-    super do |shortlist_sheet, audit_sheet|
-      lot = Lot.find(@params['lot_id'])
+  include ActionView::Helpers::NumberHelper
+  include LegalPanelForGovernment::RM6360::RatesHelper
 
-      add_supplier_details(shortlist_sheet, lot)
-      add_audit_trail(audit_sheet, lot)
+  def build
+    lot = Lot.find(@params['lot_id'])
+
+    if @params['have_you_reviewed'].nil?
+      build_with_results(lot)
+    else
+      build_with_rates(lot)
     end
   end
 
   private
+
+  def build_with_results(lot)
+    Axlsx::Package.new do |package|
+      add_supplier_details(package.workbook.add_worksheet(name: 'Supplier shortlist'), lot)
+      add_audit_trail(package.workbook.add_worksheet(name: 'Shortlist audit'), lot)
+    end
+  end
+
+  def build_with_rates(lot)
+    jurisdictions = get_jurisdictions(lot)
+    positions = get_positions(lot)
+    supplier_details_columns = get_supplier_columns(positions, lot)
+
+    Axlsx::Package.new do |package|
+      add_audit_trail(package.workbook.add_worksheet(name: 'Shortlist audit'), lot)
+
+      jurisdictions.each do |jurisdiction|
+        worksheet_name = if jurisdiction.id == 'GB'
+                           'Supplier rates'
+                         else
+                           "Supplier rates (#{jurisdiction.id})"
+                         end
+
+        add_supplier_rates(package.workbook.add_worksheet(name: worksheet_name), supplier_details_columns, lot, jurisdiction, positions)
+      end
+    end
+  end
 
   def add_audit_trail(audit_sheet, lot)
     audit_sheet.add_row ['Lot', "#{lot.number} - #{lot.name}"]
@@ -23,7 +54,8 @@ class LegalPanelForGovernment::RM6360::SupplierSpreadsheetCreator < SupplierSpre
   def add_supplier_details(shortlist_sheet, lot)
     shortlist_sheet.add_row ['Supplier name', 'Phone number', 'Email', 'Prospectus']
 
-    @supplier_frameworks.each do |supplier_framework|
+    # rubocop:disable Style/HashEachMethods
+    @supplier_frameworks.each do |supplier_framework, _rates|
       shortlist_sheet.add_row(
         [
           supplier_framework.supplier_name,
@@ -32,6 +64,23 @@ class LegalPanelForGovernment::RM6360::SupplierSpreadsheetCreator < SupplierSpre
           supplier_framework.contact_detail.additional_details["lot_#{lot.number}_prospectus_link"]
         ]
       )
+    end
+    # rubocop:enable Style/HashEachMethods
+  end
+
+  def add_supplier_rates(shortlist_sheet, supplier_details_columns, lot, jurisdiction, positions)
+    if lot.id.starts_with?('RM6360.4')
+      if @params['not_core_jurisdiction'] == 'no'
+        shortlist_sheet.add_row ['Countries:', Jurisdiction.core.order(:name).pluck(:name).join('; ')]
+      else
+        shortlist_sheet.add_row ['Country:', jurisdiction.name]
+      end
+    end
+
+    shortlist_sheet.add_row supplier_details_columns[0]
+
+    @supplier_frameworks.zip(supplier_details_columns[1..]).map do |(_supplier_framework, rates), supplier_details_column|
+      shortlist_sheet.add_row supplier_details_column + positions.map { |position_id, _position_name| display_rate(position_id, jurisdiction.id, rates) }
     end
   end
 
@@ -48,5 +97,24 @@ class LegalPanelForGovernment::RM6360::SupplierSpreadsheetCreator < SupplierSpre
                 end.order(:name).pluck(:name)
 
     sheet.add_row ['Countries', countries.join('; ')]
+  end
+
+  def get_jurisdictions(lot)
+    Jurisdiction.where(id: lot.id.starts_with?('RM6360.4') && @params['not_core_jurisdiction'] == 'yes' ? @params['jurisdiction_ids'] : ['GB']).order(:name)
+  end
+
+  def get_positions(lot)
+    lot.positions.order(:number).pluck(:id, :name)
+  end
+
+  def get_supplier_columns(positions, lot)
+    [
+      ['Supplier name', 'Prospectus'] + positions.map { |_position_id, position_name| I18n.t("legal_panel_for_government.rm6360.suppliers.rates_table.job_titles.#{position_name}") }
+    ] + @supplier_frameworks.map do |supplier_framework, _rates|
+      [
+        supplier_framework.supplier_name,
+        supplier_framework.contact_detail.additional_details["lot_#{lot.number}_prospectus_link"]
+      ]
+    end
   end
 end